Break out utility code in parser_test.go
It's fairly long and ugly and it's nice to get it out of there.
